Access FeePerKbConstants as a singleton instead of creating a new instance. Rename FeePerKbErrorCode to FeePerKbResponseCode to add the option of a successful response.
