+++ rskj-core/src/main/java/co/rsk/peg/feeperkb/FeePerKbResponseCode.java
@@ -1,0 +1,18 @@
+package co.rsk.peg.feeperkb;
+
+public enum FeePerKbResponseCode {
+    SUCCESSFUL(1),
+    GENERIC(-10),
+    NEGATIVE(-1),
+    EXCESSIVE(-2);
+
+    private final int code;
+
+    FeePerKbResponseCode(int code) {
+        this.code = code;
+    }
+
+    public int getCode() {
+        return code;
+    }
+}
